requirements.txt
annotated-types==0.5.0
anyio==3.7.1
autopep8==2.0.4
bcrypt==4.0.1
cffi==1.15.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
ecdsa==0.18.0
fastapi==0.103.1
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.4.1
pydantic_core==2.10.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
